requirements.txt
torch
transformers
diffusers
nltk
Pillow
streamlit